apps/api/pyproject.toml
[tool.poetry]
name = "np-orders-checker-api"
version = "0.1.0"
description = "FastAPI backend for the NP Orders Checker application"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11.6"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
python-multipart = "^0.0.6"
pydantic = "^2.5.0"
httpx = "^0.25.2"
alembic = "^1.17.2"
pydantic-settings = "^2.12.0"
sqlalchemy = "^2.0.0"
aiosqlite = "^0.19.0"
jose = "^1.0.0"
google-auth = "^2.46.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.0.0"
isort = "^5.12.0"

[tool.poetry.scripts]
dev = "uvicorn main:app --reload --host 0.0.0.0 --port 8000"
start = "uvicorn main:app --host 0.0.0.0 --port 8000"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11.6"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true